# RadiCorder/openapi/openapi.local.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RadiCorder | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5011/"
    }
  ],
  "paths": {
...
    "/api/recordings/{id}/tags": {
      "post": {
        "tags": [
          "RecordingsApi"
        ],
        "summary": "録音済み番組にタグを付与する",
        "operationId": "ApiAddTagsToRecording",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingTagsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/recordings/{id}/tags/{tagId}": {
      "delete": {
        "tags": [
          "RecordingsApi"
        ],
        "summary": "録音済み番組からタグを解除する",
        "operationId": "ApiRemoveTagFromRecording",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tags": {
      "get": {
        "tags": [
          "TagsApi"
        ],
        "summary": "タグ一覧を取得する",
        "operationId": "ApiGetTags",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfListOfTagEntry"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TagsApi"
        ],
        "summary": "タグを作成する",
        "operationId": "ApiCreateTag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfTagEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/tags/{id}": {
      "patch": {
        "tags": [
          "TagsApi"
        ],
        "summary": "タグを更新する",
        "operationId": "ApiUpdateTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfTagEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TagsApi"
        ],
        "summary": "タグを削除する",
        "operationId": "ApiDeleteTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/tags/merge": {
      "post": {
        "tags": [
          "TagsApi"
        ],
        "summary": "タグを統合する",
        "operationId": "ApiMergeTag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseOfEmptyData": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EmptyData"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfListOfTagEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/TagEntry"
            }
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfTagEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TagEntry"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EmptyData": {
        "type": "object"
      },
...
      "RecordingTagsRequest": {
        "type": "object",
        "properties": {
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "TagEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "recordingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastUsedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TagMergeRequest": {
        "type": "object",
        "properties": {
          "fromTagId": {
            "type": "string",
            "format": "uuid"
          },
          "toTagId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "TagUpsertRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}